do not fail price calculation if product has no extras
For some reason, products without extras return an empty array
for the ingredient_extras_with_details attribute, instead of
an Hash. This of course breaks the price calculation.

Instead of actually checking for that, we can simply return
early if the user didn't select any extra ingredients. If they
had, but this product actually doesn't have extras, then it will
still fail… which is good I guess?